requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
openai==1.3.7
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
matplotlib==3.8.2
openpyxl==3.1.2
